Clean up the MPT and Falcon model (#10)
In this PR we

Split the Falcon model to two: inference from HF model and inference with mlflow to make it as the same format as MPT model
Change MPT model to run on MLR 13.1

---------
