+++ llm-examples/mpt/mpt-7b/01_load_inference.py
@@ -3,8 +3,20 @@
 # MAGIC # Load and Inference MPT-7B model on Databricks
 # MAGIC
 # MAGIC Environment for this notebook:
-# MAGIC - Runtime: 13.2 GPU ML Runtime
+# MAGIC - Runtime: 13.1 GPU ML Runtime
 # MAGIC - Instance: `g5.4xlarge` on AWS
+
+# COMMAND ----------
+
+# Skip this step if running on Databricks runtime 13.2 GPU and above.
+!wget -O /local_disk0/tmp/libcusparse-dev-11-7_11.7.3.50-1_amd64.deb https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2204/x86_64/libcusparse-dev-11-7_11.7.3.50-1_amd64.deb && \
+  dpkg -i /local_disk0/tmp/libcusparse-dev-11-7_11.7.3.50-1_amd64.deb && \
+  wget -O /local_disk0/tmp/libcublas-dev-11-7_11.10.1.25-1_amd64.deb https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2204/x86_64/libcublas-dev-11-7_11.10.1.25-1_amd64.deb && \
+  dpkg -i /local_disk0/tmp/libcublas-dev-11-7_11.10.1.25-1_amd64.deb && \
+  wget -O /local_disk0/tmp/libcusolver-dev-11-7_11.4.0.1-1_amd64.deb https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2204/x86_64/libcusolver-dev-11-7_11.4.0.1-1_amd64.deb && \
+  dpkg -i /local_disk0/tmp/libcusolver-dev-11-7_11.4.0.1-1_amd64.deb && \
+  wget -O /local_disk0/tmp/libcurand-11-7_10.2.10.91-1_amd64.deb https://developer.download.nvidia.com/compute/cuda/repos/ubuntu2204/x86_64/libcurand-11-7_10.2.10.91-1_amd64.deb && \
+  dpkg -i /local_disk0/tmp/libcurand-11-7_10.2.10.91-1_amd64.deb
 
 # COMMAND ----------
 
@@ -32,6 +44,8 @@ model = transformers.AutoModelForCausalLM.from_pretrained(
   trust_remote_code=True,
   cache_dir="/local_disk0/.cache/huggingface/"
 )
+
+model.eval()
 
 tokenizer = transformers.AutoTokenizer.from_pretrained("EleutherAI/gpt-neox-20b", padding_side="left")
 
@@ -73,5 +87,22 @@ print(generate_text(question))
 
 # COMMAND ----------
 
+generate_kwargs = {
+  'max_new_tokens': 100,
+  'temperature': 0.2,
+  'top_p': 0.9,
+  'top_k': 50,
+  'repetition_penalty': 1.0,
+  'no_repeat_ngram_size': 0,
+  'use_cache': True,
+  'do_sample': True,
+  'eos_token_id': tokenizer.eos_token_id,
+  'pad_token_id': tokenizer.eos_token_id,
+}
+
 questions = ["what is ML?", "Name 10 colors."]
-print(generate_text(questions))
+print(generate_text(questions, **generate_kwargs))
+
+# COMMAND ----------
+
+
